+++ packages/block-editor/src/components/inspector-controls-tabs/settings-tab.js
@@ -1,43 +1,18 @@
 /**
- * WordPress dependencies
- */
-import { __experimentalUseSlotFills as useSlotFills } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-
-/**
  * Internal dependencies
  */
 import AdvancedControls from './advanced-controls-panel';
-import InspectorControlsGroups from '../inspector-controls/groups';
-import {
-	default as InspectorControls,
-	InspectorAdvancedControls,
-} from '../inspector-controls';
+import { default as InspectorControls } from '../inspector-controls';
 
-const SettingsTab = ( { hasSingleBlockSelection = false } ) => {
-	const { default: defaultGroup } = InspectorControlsGroups;
-	const settingsFills = [
-		...( useSlotFills( defaultGroup.Slot.__unstableName ) || [] ),
-		...( useSlotFills( InspectorAdvancedControls.slotName ) || [] ),
-	];
-
-	return (
-		<>
-			<InspectorControls.Slot />
-			{ hasSingleBlockSelection && (
-				<div>
-					<AdvancedControls />
-				</div>
-			) }
-			{ ! settingsFills.length && (
-				<span className="block-editor-block-inspector__no-block-tools">
-					{ hasSingleBlockSelection
-						? __( 'This block has no settings.' )
-						: __( 'The selected blocks have no settings.' ) }
-				</span>
-			) }
-		</>
-	);
-};
+const SettingsTab = ( { showAdvancedControls = false } ) => (
+	<>
+		<InspectorControls.Slot />
+		{ showAdvancedControls && (
+			<div>
+				<AdvancedControls />
+			</div>
+		) }
+	</>
+);
 
 export default SettingsTab;